# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trees API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://arboraid.co/beta/public"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/wishlist/add/{tree}": {
            "post": {
                "tags": [
                    "WishList"
                ],
                "summary": "Add a tree to the wish list",
                "operationId": "0ef0d94ee4397820b90e08ec7f60a992",
                "parameters": [
                    {
                        "name": "tree",
                        "in": "path",
                        "description": "ID of the tree to add",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tree added to wish list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tree added to wish list!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/wishlist/remove/{tree}": {
            "delete": {
                "tags": [
                    "WishList"
                ],
                "summary": "Remove a tree from the wish list",
                "operationId": "d863665789867cd880d7c6a4588083b8",
                "parameters": [
                    {
                        "name": "tree",
                        "in": "path",
                        "description": "ID of the tree to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tree removed from wish list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tree removed from wish list!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/wishlist": {
            "get": {
                "tags": [
                    "WishList"
                ],
                "summary": "Get all trees in the wish list",
                "operationId": "6a9e5f3a362662290f5bea9a4ed4ef1b",
                "responses": {
                    "200": {
                        "description": "List of trees in the wish list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Enter 'Bearer {token}' to access the endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}